requirements.txt
streamlit>=1.24.0
python-dotenv
groq
pydantic